feature: order service kafka integration complete

+++ bookingservice/src/main/java/com/bookings/bookingservice/request/BookingRequest.java
@@ -15,5 +15,5 @@ import lombok.NoArgsConstructor;
 public class BookingRequest {
 	private Long userId;
 	private Long eventId;
-	private int ticketCount;
+	private Long ticketCount;
 }
